+++ example/lib/cupertino_example.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:catcher/catcher_plugin.dart';
+import 'package:catcher/catcher.dart';
 
 main() {
   CatcherOptions debugOptions = CatcherOptions(DialogReportMode(), [