+++ example/lib/custom_report_mode_example.dart
@@ -1,6 +1,6 @@
 import 'package:catcher/model/platform_type.dart';
 import 'package:flutter/material.dart';
-import 'package:catcher/catcher_plugin.dart';
+import 'package:catcher/catcher.dart';
 
 main() {
   CatcherOptions debugOptions = CatcherOptions(CustomPageReportMode(), [