+++ example/lib/file_example.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:catcher/mode/dialog_report_mode.dart';
 import 'package:catcher/model/catcher_options.dart';
 import 'package:flutter/material.dart';
-import 'package:catcher/catcher_plugin.dart';
+import 'package:catcher/catcher.dart';
 import 'package:permission_handler/permission_handler.dart';
 
 void main() async {