+++ example/lib/localization_example.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:catcher/catcher_plugin.dart';
+import 'package:catcher/catcher.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
 main() {