+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.14
+* General refactor
+* [BREAKING_CHANGE] You should use import "import 'package:catcher/catcher.dart';" instead import 'package:catcher/catcher_plugin.dart';
+
 ## 0.3.13
 * Fixed pub dev score error